fix a few single-instance clippy warnings
all of these warnings are only emitted once, so just put them into this
commit.

- clippy::unit_arg
- clippy::let_unit_value
- clippy::empty_line_after_doc_comments
- clippy::unused_unit
- clippy::redundant_field_names
- clippy::len_without_is_empty
- clippy::unnecessary_cast

+++ crates/steel-core/src/primitives/vectors.rs
@@ -949,7 +949,7 @@ pub fn mut_vec_set(vec: &HeapRef<Vec<SteelVal>>, i: usize, value: SteelVal) -> R
     }
 
     // Update the vector position
-    guard[i as usize] = value;
+    guard[i] = value;
 
     Ok(SteelVal::Void)
 }
